Add Monster

+++ Portfolio_Game/Source/Header/FbxLoader.h
@@ -37,22 +37,22 @@ public:
 	FbxLoader();
 	~FbxLoader();
 
-	HRESULT LoadFBX(std::vector<SkinnedVertex>& outVertexVector, std::vector<uint16_t>& outIndexVector, SkinnedData& outSkinnedData, const std::string& ClipName, std::vector<Material>& outMaterial, std::string fileName);
+	HRESULT LoadFBX(std::vector<SkinnedVertex>& outVertexVector, std::vector<uint32_t>& outIndexVector, SkinnedData& outSkinnedData, const std::string& ClipName, std::vector<Material>& outMaterial, std::string fileName);
 	HRESULT LoadFBX(SkinnedData& outSkinnedData, const std::string& clipName, std::string fileName);
-	bool LoadTXT(std::vector<SkinnedVertex>& outVertexVector, std::vector<uint16_t>& outIndexVector, SkinnedData& outSkinnedData, const std::string& clipName, std::vector<Material>& outMaterial, std::string fileName);
+	bool LoadTXT(std::vector<SkinnedVertex>& outVertexVector, std::vector<uint32_t>& outIndexVector, SkinnedData& outSkinnedData, const std::string& clipName, std::vector<Material>& outMaterial, std::string fileName);
 	bool LoadAnimation(AnimationClip& animation, const std::string& clipName, std::string fileName);
 
 	void GetSkeletonHierarchy(FbxNode * pNode, SkinnedData& outSkinnedData, int curIndex, int parentIndex);
 	void GetControlPoints(fbxsdk::FbxNode * pFbxRootNode);
 	void GetAnimation(FbxScene * pFbxScene, FbxNode * pFbxChildNode, std::string & outAnimationName, const std::string& ClipName);
 	void GetOnlyAnimation(FbxScene* pFbxScene, FbxNode * pFbxChildNode, AnimationClip& animation, const std::string clipName);
-	void GetVerticesAndIndice(fbxsdk::FbxMesh * pMesh, std::vector<SkinnedVertex> & outVertexVector, std::vector<uint16_t> & outIndexVector, SkinnedData& outSkinnedData);
+	void GetVerticesAndIndice(fbxsdk::FbxMesh * pMesh, std::vector<SkinnedVertex> & outVertexVector, std::vector<uint32_t> & outIndexVector, SkinnedData& outSkinnedData);
 	void GetMaterials(FbxNode * pNode, std::vector<Material>& outMaterial);
 	void GetMaterialAttribute(FbxSurfaceMaterial * pMaterial, std::vector<Material>& outMaterial);
 	void GetMaterialTexture(FbxSurfaceMaterial * pMaterial, Material & Mat);
 	FbxAMatrix GetGeometryTransformation(FbxNode * pNode);
 
-	void ExportFBX(std::vector<SkinnedVertex>& outVertexVector, std::vector<uint16_t>& outIndexVector, SkinnedData& outSkinnedData, const std::string& clipName, std::vector<Material>& outMaterial, std::string fileName);
+	void ExportFBX(std::vector<SkinnedVertex>& outVertexVector, std::vector<uint32_t>& outIndexVector, SkinnedData& outSkinnedData, const std::string& clipName, std::vector<Material>& outMaterial, std::string fileName);
 	void ExportAnimation(const AnimationClip& animation, std::string fileName, const std::string& clipName);
 
 private:
